+++ pyproject.toml
@@ -29,9 +29,21 @@ repository = "https://github.com/zerospeech/vocolab"
 voco = "vocolab.admin.main:run_cli"
 
 [project.optional-dependencies]
-extend = [
+
+extras = [
     # todo migrate towards normal install when package is published
-    "git+ssh://git@github.com/zerospeech/vocolab-ext.git"
+    "vocolab-ext @ https://github.com/zerospeech/vocolab-ext/archive/master.zip"
+]
+
+eval = [
+    "vocolab[extras]",
+    "celery",
+]
+
+zerospeech = [
+    "vocolab[extras]",
+    # todo migrate this to pip version
+    "zerospeech-benchmarks @ https://github.com/zerospeech/vocolab-ext/archive/develop.zip"
 ]
 
 dev = [
@@ -42,7 +54,6 @@ dev = [
     "build",
     "icecream"
 ]
-
 
 
 [build-system]
