+++ vocolab/api/endpoints/leaderboards.py
@@ -1,12 +1,18 @@
 """ Routing for /leaderboards section of the API
 This section handles leaderboard data
 """
+import tempfile
+from pathlib import Path
 
 from fastapi import (
-    APIRouter
+    APIRouter, BackgroundTasks
 )
+from fastapi.responses import FileResponse
+from starlette.background import BackgroundTask
 
-from vocolab.data import models, model_queries
+from vocolab_ext import leaderboards as leaderboard_ext
+
+from vocolab.data import model_queries
 from vocolab.settings import get_settings
 
 router = APIRouter()
@@ -14,38 +20,46 @@ _settings = get_settings()
 
 
 @router.get("/list")
-async def get_list():
-    pass
+async def get_list() -> list[str]:
+    ld_list = await model_queries.LeaderboardList.get_all()
+    return [
+        ld.label for ld in ld_list
+    ]
 
 
-@router.get('{leaderboard_id}/info')
-async def get_leaderboard_info(leaderboard_id: str):
+@router.get('{leaderboard}/info')
+async def get_leaderboard_info(leaderboard: str):
     """ Return information of a specific challenge """
-    return await model_queries.Leaderboard.get(leaderboard_id=leaderboard_id)
+    return await model_queries.Leaderboard.get(leaderboard_id=leaderboard)
 
 
-@router.get("{leaderboard_id}/json")
-async def get_leaderboard_entries_as_json(leaderboard_id: int):
-    pass
-    # try:
-    #     leaderboard = await leaderboardQ.get_leaderboard(leaderboard_id=leaderboard_id)
-    # except ValueError:
-    #     raise exc.ResourceRequestedNotFound(f'No leaderboard with id {leaderboard_id}')
-    #
-    # if leaderboard.path_to.is_file():
-    #     return api_lib.file2dict(leaderboard.path_to)
-    # else:
-    #     return dict(
-    #         updatedOn=datetime.now().isoformat(),
-    #         data=[]
-    #     )
+@router.get("{leaderboard}/json")
+async def get_leaderboard_entries_as_json(leaderboard: str):
+    """ Return a leaderboard into a json format """
+    entry_list = await model_queries.LeaderboardEntryList.get_from_leaderboard(leaderboard)
+    return entry_list.as_leaderboard()
 
 
-@router.get("{leaderboard_id}/csv")
-async def get_leaderboard_entries_as_csv(leaderboard_id: int):
-    pass
+@router.get("{leaderboard}/csv")
+async def get_leaderboard_entries_as_csv(leaderboard: str):
+    def clean(file: tempfile.NamedTemporaryFile):
+        """ clean temp file """
+        Path(file.name).unlink(missing_ok=True)
+
+    # load objects
+    entry_list = await model_queries.LeaderboardEntryList.get_from_leaderboard(leaderboard)
+    ld_mngr = leaderboard_ext.LeaderboardManager.load_leaderboard_from_obj(leaderboard, entry_list.as_leaderboard())
+
+    # Write csv into tmp file
+    tmp_file = tempfile.NamedTemporaryFile(suffix='.csv', delete=False)
+    ld_mngr.export_as_csv(file=Path(tmp_file.name))
+
+    # return file w/ clean-up bg-task
+    return FileResponse(tmp_file.name, background=BackgroundTask(clean, file=tmp_file))
 
 
-@router.get("{leaderboard_id}/entry/{entry_id}")
-async def get_leaderboard_entry(leaderboard_id: int, entry_id: str):
-    pass
+@router.get("{leaderboard}/entry/{entry_id}")
+async def get_leaderboard_entry(leaderboard: str, entry_id: str):
+    entry = await model_queries.LeaderboardEntry.get(entry_id)
+    assert entry.leaderboard_id == leaderboard
+    return entry.data
