+++ vocolab/data/model_queries/__init__.py
@@ -1,3 +1,4 @@
 from .auth import *
 from .challenges import *
 from .models import *
+from .leaderboard import *
